Lay out index page
Separate footer/header for post vs index pages

This lets us have slightly different grids for posts and pages. Pages
need slightly wider left columns to accomodate putting content there.
But I need to have the columns in the header and footer line up with the
content.

Article list is two colummns, date vs title. This is styled with article
tag and asides for the date column.

The ephemera article listing is one column, but uses the article tag as
well.